chore: move image.py to api/
